He hecho administradorServiceTest con Mockito. Presenta un problema: las funciones que devuelven void como save y delete a Mockito no le gustan. COn esto supongo que para esas pruebas no se puede usar Mockito (preguntar).
